Add Column QTY To Product Table

+++ database/migrations/2022_05_12_015110_create_tbl_order.php
@@ -18,7 +18,7 @@ class CreateTblOrder extends Migration
             $table->integer('customer_id');
             $table->integer('shipping_id');
             $table->integer('payment_id');
-            $table->float('order_total');
+            $table->double('order_total');
             $table->integer('order_status');
             $table->timestamps();
         });
